requirements.txt
openai
streamlit
requests
python-dotenv

